+ .NET: - downgrade Microsoft.Extensions to versions tested with .NETCore - update System.CommandLine to beta - remove unused + WIP
