+++ src/Todo.CLI/Commands/ExportCommand.cs
@@ -8,16 +8,19 @@ namespace MSTTool.Commands
 {
     public class ExportCommand : Command
     {
-        public ExportCommand(IServiceProvider serviceProvider) : base("list")
+        public ExportCommand(IServiceProvider serviceProvider) : base("export")
         {
             Description = "Exports ToDo items to JSON files";
 
+            /*fnord - specific list, all
             AddOption(GetAllOption());
             AddOption(GetNoStatusOption());
+            */
 
-            Handler = ListCommandHandler.Create(serviceProvider);
+            //Handler = ListCommandHandler.Create(serviceProvider);
         }
 
+        /*fnord
         private Option GetAllOption()
         {
             return new Option(new string[] { "-a", "--all" }, "Lists all to do items including the completed ones.");
@@ -27,6 +30,7 @@ namespace MSTTool.Commands
         {
             return new Option(new string[] { "--no-status" }, "Suppresses the bullet indicating whether the item is completed or not.");
         }
+        */
 
     }
 }
